implemented lifecycle-management for Android
implemented support for lifecycle management for the Android-side. The NativeDeviceOrientationCommunicator now has a pause and a resume method which pauses and resumes the active orientationlistener, if any. User's will need to call the pause and resume methods if they use the NativeDeviceOrientationCommunicator directly.
If they use the NativeDeviceOrientationReader it is handled for them.