lifecycle-support, useSensor-support iOS (#8)
* implemented lifecycle-management for Android

implemented support for lifecycle management for the Android-side. The NativeDeviceOrientationCommunicator now has a pause and a resume method which pauses and resumes the active orientationlistener, if any. User's will need to call the pause and resume methods if they use the NativeDeviceOrientationCommunicator directly.
If they use the NativeDeviceOrientationReader it is handled for them.

* - added pause and resume methods to the iOS-version. They still need to be implemented.

* useSensor-support for iOS

added back the useSensor-functionality for the iOS-version

* Update NativeDeviceOrientationPlugin.m

created seperate methods for starting and stopping the observing of orientation changes. This makes it easier to implement the pause and resume functionality later on.

* improved code-structure

- seperated the logic of the NativeDeviceOrientationPlugin into seperate classes. It is setup similar to the Android version. There is a IOrientationListener-protocol which is implemented by the OrientationListener and the SensorListener.
- implemented pause -method. The resume-method isn't finished yet.
- implemented useSensor support for the getOrientation-method for iOS

* implemented resume, code clean-up

- implemented the resume method
- code clean-up, removed old code
- removed logging